requirements.txt
absl-py==0.14.0
astor==0.8.1
cached-property==1.5.2
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
fashionpedia==1.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
imageio==2.9.0
importlib-metadata==4.8.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
pandas==0.25.3
Pillow==8.3.2
protobuf==3.18.0
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2020.9.3
tqdm==4.62.3
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wget==3.2
wrapt==1.12.1
zipp==3.6.0
